+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added List detail view. [#155](https://github.com/verse-pbc/issues/issues/155)
 - Added view for managing users in a list. [#135](https://github.com/verse-pbc/issues/issues/135)
 - Added ability to delete lists. [#136](https://github.com/verse-pbc/issues/issues/136)
+- Added analytics for feed source selection and lists. [#129](https://github.com/verse-pbc/issues/issues/129)
 
 ### Internal Changes
 - Added function for creating a new list and a test verifying list editing. [#112](https://github.com/verse-pbc/issues/issues/112)